Split the components into separate modules
