+++ mirrormanager2/templates/fedora/host_category_url_new.html
@@ -8,13 +8,13 @@
 <h2>Add host category URL</h2>
 
 <p>
-  Back to <a href="{{ url_for('site_view', site_id=host.site_id) }}">
-    {{ host.site.name }}</a> / <a href="{{ url_for('host_view', host_id=host.id) }}">
-    {{ host.name }}</a> / <a href="{{ url_for('host_category', host_id=host.id,
+  Back to <a href="{{ url_for('base.site_view', site_id=host.site_id) }}">
+    {{ host.site.name }}</a> / <a href="{{ url_for('base.host_view', host_id=host.id) }}">
+    {{ host.name }}</a> / <a href="{{ url_for('base.host_category', host_id=host.id,
     hc_id=hostcategory.id) }}">{{ hostcategory.category.name }} </a>
 </p>
 
-<form action="{{ url_for('host_category_url_new',
+<form action="{{ url_for('base.host_category_url_new',
                  host_id=host.id, hc_id=hostcategory.id) }}" method="POST">
   <table>
     {{ render_field_in_row(form.url, after='URL (rsync, https, http) pointing to the top directory') }}
@@ -24,7 +24,7 @@
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Create">
-    <a href="{{ url_for('host_view', host_id=host.id) }}">
+    <a href="{{ url_for('base.host_view', host_id=host.id) }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
