+++ mirrormanager2/templates/fedora/host_country_new.html
@@ -8,7 +8,7 @@
 <h2>Add host country allowed</h2>
 
 <p>
-  Back to <a href="{{ url_for('host_view', host_id=host.id) }}">
+  Back to <a href="{{ url_for('base.host_view', host_id=host.id) }}">
     {{ host.name }}</a>
 </p>
 
@@ -19,13 +19,13 @@
   honor this.
 </p>
 
-<form action="{{ url_for('host_country_new', host_id=host.id) }}" method="POST">
+<form action="{{ url_for('base.host_country_new', host_id=host.id) }}" method="POST">
   <table>
     {{ render_field_in_row(form.country) }}
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Create">
-    <a href="{{ url_for('host_view', host_id=host.id) }}">
+    <a href="{{ url_for('base.host_view', host_id=host.id) }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
