+++ mirrormanager2/templates/fedora/host_netblock_new.html
@@ -8,7 +8,7 @@
 <h2>Add host netblock</h2>
 
 <p>
-  Back to <a href="{{ url_for('host_view', host_id=host.id) }}">
+  Back to <a href="{{ url_for('base.host_view', host_id=host.id) }}">
     {{ host.name }}</a>
 </p>
 
@@ -21,14 +21,14 @@
   (no RFC1918 private space addresses).
 </p>
 
-<form action="{{ url_for('host_netblock_new', host_id=host.id) }}" method="POST">
+<form action="{{ url_for('base.host_netblock_new', host_id=host.id) }}" method="POST">
   <table>
     {{ render_field_in_row(form.netblock) }}
     {{ render_field_in_row(form.name) }}
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Create">
-    <a href="{{ url_for('host_view', host_id=host.id) }}">
+    <a href="{{ url_for('base.host_view', host_id=host.id) }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
