+++ mirrormanager2/templates/fedora/index.html
@@ -45,7 +45,7 @@ networks globally.
 {% for product in products %}
 <tr class="matrix_section">
   <td class="matrix_section" rowspan="{{product.displayed_versions | length }}">
-    <a href="{{ url_for('list_mirrors', p_name=product.name) }}">
+    <a href="{{ url_for('base.list_mirrors', p_name=product.name) }}">
       {{product.name}}
     </a>
   </td>
@@ -58,7 +58,7 @@ networks globally.
       {% endif %}
     {% endif %}
     <td>
-      <a href="{{ url_for('list_mirrors',
+      <a href="{{ url_for('base.list_mirrors',
                 p_name=product.name, p_version=version.name) }}">
         {{ version.display_name or version.name }}
       </a>
@@ -66,7 +66,7 @@ networks globally.
     {% for arch in arches %}
       {% if arch in version.arches %}
         <td>
-          <a href="{{ url_for('list_mirrors',
+          <a href="{{ url_for('base.list_mirrors',
                     p_name=product.name, p_version=version.name, p_arch=arch) }}">
             {{ arch }}
           </a>
