+++ mirrormanager2/templates/fedora/login.html
@@ -7,7 +7,7 @@
 {% block content %}
 <h2>Login</h2>
 
-<form action="{{ url_for('do_login') }}" method="POST">
+<form action="{{ url_for('local_auth.do_login') }}" method="POST">
   <input type="hidden" value="{{ next_url }}" name="next_url" />
   <table>
     {{ render_field_in_row(form.username) }}
@@ -15,7 +15,7 @@
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Login">
-    <a href="{{ url_for('index') }}">
+    <a href="{{ url_for('base.index') }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
@@ -23,12 +23,12 @@
 </form>
 
 <p>
-  If you do not have already an account: <a href="{{ url_for('new_user') }}">
+  If you do not have already an account: <a href="{{ url_for('local_auth.new_user') }}">
   Create one!</a>
 </p>
 
 <p>
-  <a href="{{ url_for('lost_password') }}">
+  <a href="{{ url_for('local_auth.lost_password') }}">
   Forgot your password?</a>
 </p>
