+++ mirrormanager2/templates/fedora/maps.html
@@ -6,8 +6,8 @@
 {% block content %}
 <h1>Maps</h1>
 {%- if config['SHOW_MAPS'] -%}
-Map of all public mirrors - <a href="{{url_for('index')}}{{ config['INTERACTIVE_MAP'] }}">Interactive Map</a>
-<img src="{{url_for('index')}}{{ config['STATIC_MAP'] }}" border="0" alt="alt" class="centered"/>
+Map of all public mirrors - <a href="{{url_for('base.index')}}{{ config['INTERACTIVE_MAP'] }}">Interactive Map</a>
+<img src="{{url_for('base.index')}}{{ config['STATIC_MAP'] }}" border="0" alt="alt" class="centered"/>
 {%- else -%}
 Actually, this is disabled and you should not be here. Go away!
 {%- endif -%}
