+++ mirrormanager2/templates/fedora/password_change.html
@@ -7,13 +7,13 @@
 {% block content %}
 <h2>Lost password?</h2>
 
-<form action="{{ url_for('lost_password') }}" method="POST">
+<form action="{{ url_for('local_auth.lost_password') }}" method="POST">
   <table>
     {{ render_field_in_row(form.username) }}
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Send email">
-    <a href="{{ url_for('index') }}">
+    <a href="{{ url_for('base.index') }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
