+++ mirrormanager2/templates/fedora/password_reset.html
@@ -7,14 +7,14 @@
 {% block content %}
 <h2>Change password</h2>
 
-<form action="{{ url_for('reset_password', token=token) }}" method="POST">
+<form action="{{ url_for('local_auth.reset_password', token=token) }}" method="POST">
   <table>
     {{ render_field_in_row(form.password) }}
     {{ render_field_in_row(form.confirm_password) }}
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Update">
-    <a href="{{ url_for('index') }}">
+    <a href="{{ url_for('base.index') }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
