+++ mirrormanager2/templates/fedora/site_new.html
@@ -54,7 +54,7 @@
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Create">
-    <a href="{{ url_for('index') }}">
+    <a href="{{ url_for('base.index') }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
