+++ mirrormanager2/templates/fedora/user_new.html
@@ -13,7 +13,7 @@
   You will be asked to provide a valid email address.
 </p>
 
-<form action="{{ url_for('new_user') }}" method="POST">
+<form action="{{ url_for('base.new_user') }}" method="POST">
   <table>
     {{ render_field_in_row(form.user_name) }}
     {{ render_field_in_row(form.display_name) }}
@@ -23,7 +23,7 @@
   </table>
   <p class="buttons indent">
     <input type="submit" class="submit positive button" value="Create user">
-    <a href="{{ url_for('index') }}">
+    <a href="{{ url_for('base.index') }}">
       <input type="button" value="Cancel" class="button">
     </a>
     {{ form.csrf_token }}
